Can now set remote server
An arbitrary IP address and port can now be input by the user, enabling
SMS to be forwarded anywhere. Not sure if this will make it to the final
product as the vision is to have SMS between the phone and a PC on a
LAN, but we will see. The commit still outlines the changes to re-add
the feature if it is ever removed.